move desugaring to other module
